+++ pirds_library/PIRDS.h
@@ -92,21 +92,21 @@ typedef struct Message
 
 /* Fill the byte buffer with a PIRDS-standard bytes from the
    Measurement Object */
-int fill_byte_buffer_measurement(Measurement* m,uint8_t* buff,uint16_t blim);
+uint16_t fill_byte_buffer_measurement(Measurement* m,uint8_t* buff,uint16_t blim);
 
 Measurement get_measurement_from_buffer(uint8_t* buff,uint16_t blim);
 
-int fill_JSON_buffer_measurement(Measurement* m,char* buff,uint16_t blim);
+uint16_t fill_JSON_buffer_measurement(Measurement* m,char* buff,uint16_t blim);
 
 Measurement get_measurement_from_JSON(char* buff,uint16_t blim);
 
 
 
-int fill_byte_buffer_message(Message* m,uint8_t* buff,uint16_t blim);
+uint16_t fill_byte_buffer_message(Message* m,uint8_t* buff,uint16_t blim);
 
 Message get_message_from_buffer(uint8_t* buff,uint16_t blim);
 
-int fill_JSON_buffer_message(Message* m,char* buff,uint16_t blim);
+uint16_t fill_JSON_buffer_message(Message* m,char* buff,uint16_t blim);
 
 Message get_message_from_JSON(char* buff,uint16_t blim);
